+++ go/cauldron/worker/activities/sparkhttp/BUILD.bazel
@@ -1,0 +1,21 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "module.go",
+        "spark_activities.go",
+    ],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/cauldron/worker/activities/sparkhttp",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//go/cauldron/worker/compute/spark:go_default_library",
+        "//go/cauldron/worker/http:go_default_library",
+        "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
+        "@com_github_cadence_workflow_starlark_worker//workflow:go_default_library",
+        "@io_k8s_sigs_controller_runtime//pkg/log:go_default_library",
+        "@org_uber_go_fx//:go_default_library",
+        "@org_uber_go_yarpc//yarpcerrors:go_default_library",
+        "@org_uber_go_zap//:go_default_library",
+    ],
+)
