+++ go/cauldron/worker/compute/ray/BUILD.bazel
@@ -3,9 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "sparkone.go",
+        "raycluster.go",
+        "rayjob.go",
     ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/spark",
+    importpath = "github.com/michelangelo-ai/michelangelo/go/cauldron/worker/compute/ray",
     visibility = ["//visibility:public"],
     deps = [
         "@io_k8s_api//core/v1:go_default_library",
