+++ go/cauldron/worker/compute/ray/raycluster.go
