+++ go/cauldron/worker/compute/ray/rayjob.go
