+++ go/cauldron/worker/compute/spark/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["sparkone.go"],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/cauldron/worker/compute/spark",
+    visibility = ["//visibility:public"],
+    deps = ["@io_k8s_apimachinery//pkg/apis/meta/v1/unstructured:go_default_library"],
+)
