+++ go/cauldron/worker/compute/spark/sparkone.go
