+++ go/cauldron/worker/http/BUILD.bazel
@@ -2,10 +2,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "config.go",
-    ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/activities/http",
+    srcs = ["config.go"],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/cauldron/worker/http",
     visibility = ["//visibility:public"],
     deps = [
         "@org_uber_go_fx//:go_default_library",
