+++ go/cauldron/worker/plugins/rayhttp/BUILD.bazel
@@ -6,12 +6,12 @@ go_library(
         "plugin.go",
         "starlark_module.go",
     ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/plugins/sparkhttp",
+    importpath = "github.com/michelangelo-ai/michelangelo/go/cauldron/worker/plugins/rayhttp",
     visibility = ["//visibility:public"],
     deps = [
-        "//go/worker/activities/sparkhttp:go_default_library",
+        "//go/cauldron/worker/activities/rayhttp:go_default_library",
+        "//go/cauldron/worker/compute/ray:go_default_library",
         "//go/worker/plugins/utils:go_default_library",
-        "//go/worker/spark:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//ext:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//service:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
