+++ go/cauldron/worker/plugins/rayhttp/plugin.go
