+++ go/cauldron/worker/plugins/sparkhttp/plugin.go
