+++ go/cmd/controllermgr/BUILD.bazel
@@ -11,7 +11,6 @@ go_library(
         "//go/api/handler:go_default_library",
         "//go/base/config:go_default_library",
         "//go/base/env:go_default_library",
-        "//go/base/workflowclient/cadenceclient:go_default_library",
         "//go/base/workflowclient/temporalclient:go_default_library",
         "//go/base/zapfx:go_default_library",
         "//go/components/pipeline:go_default_library",
