+++ go/cmd/worker/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
         "//go/base/config:go_default_library",
         "//go/base/env:go_default_library",
         "//go/base/zapfx:go_default_library",
+        "//go/cauldron/worker:go_default_library",
         "//go/worker:go_default_library",
         "@org_uber_go_fx//:go_default_library",
     ],
