+++ go/worker/activities/module.go
@@ -2,8 +2,8 @@ package activities
 
 import (
 	"github.com/michelangelo-ai/michelangelo/go/worker/activities/cachedoutput"
-	"github.com/michelangelo-ai/michelangelo/go/worker/activities/rayhttp"
-	"github.com/michelangelo-ai/michelangelo/go/worker/activities/sparkhttp"
+	"github.com/michelangelo-ai/michelangelo/go/cauldron/worker/activities/rayhttp"
+	"github.com/michelangelo-ai/michelangelo/go/cauldron/worker/activities/sparkhttp"
 	"github.com/michelangelo-ai/michelangelo/go/worker/activities/storage"
 	"go.uber.org/fx"
 )
