--- go/worker/ray/BUILD.bazel
@@ -1,15 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "raycluster.go",
-        "rayjob.go",
-    ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/ray",
-    visibility = ["//visibility:public"],
-    deps = [
-        "@io_k8s_api//core/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/v1/unstructured:go_default_library",
-    ],
-)
